+++ src/pages/Adjustments.tsx
@@ -1,4 +1,4 @@
-import { Layout } from "./components/template/Layout";
+import { Layout } from "../components/template/Layout";
 
 
 export default function Adjustments() {
